requirements.txt
# Headless Sentinel - Python Dependencies

# Core dependencies
pywinrm>=0.4.3
duckdb>=0.9.2
pandas>=2.0.0
pyyaml>=6.0
keyring>=24.0.0
click>=8.1.0
rich>=13.0.0
aiohttp>=3.9.0

# Additional utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Optional: For enhanced security
cryptography>=41.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Note: This application is designed for Windows environments
# Ensure you have WinRM enabled on target machines
